Refactor users service and reorganize backup structure
