+++ users/api/http/backup/responses.go
@@ -1,0 +1,42 @@
+package backup
+
+import "net/http"
+
+type backupUserRes struct {
+	ID       string         `json:"id"`
+	Email    string         `json:"email"`
+	Password string         `json:"password"`
+	Metadata map[string]any `json:"metadata,omitempty"`
+	Status   string         `json:"status"`
+}
+
+type backupRes struct {
+	Users []backupUserRes `json:"users"`
+	Admin backupUserRes   `json:"admin"`
+}
+
+func (res backupRes) Code() int {
+	return http.StatusOK
+}
+
+func (res backupRes) Headers() map[string]string {
+	return map[string]string{}
+}
+
+func (res backupRes) Empty() bool {
+	return false
+}
+
+type restoreRes struct{}
+
+func (res restoreRes) Code() int {
+	return http.StatusCreated
+}
+
+func (res restoreRes) Headers() map[string]string {
+	return map[string]string{}
+}
+
+func (res restoreRes) Empty() bool {
+	return true
+}
