+++ users/api/http/backup/transport.go
@@ -1,0 +1,97 @@
+// Copyright (c) Mainflux
+// SPDX-License-Identifier: Apache-2.0
+
+package backup
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+	"strings"
+
+	"github.com/MainfluxLabs/mainflux/logger"
+	"github.com/MainfluxLabs/mainflux/pkg/apiutil"
+	"github.com/MainfluxLabs/mainflux/pkg/errors"
+	"github.com/MainfluxLabs/mainflux/pkg/uuid"
+	"github.com/MainfluxLabs/mainflux/users"
+	kitot "github.com/go-kit/kit/tracing/opentracing"
+	kithttp "github.com/go-kit/kit/transport/http"
+	"github.com/go-zoo/bone"
+	"github.com/opentracing/opentracing-go"
+)
+
+// MakeHandler returns a HTTP handler for API endpoints.
+func MakeHandler(svc users.Service, mux *bone.Mux, tracer opentracing.Tracer, logger logger.Logger) *bone.Mux {
+	opts := []kithttp.ServerOption{
+		kithttp.ServerErrorEncoder(apiutil.LoggingErrorEncoder(logger, encodeError)),
+	}
+
+	mux.Get("/backup", kithttp.NewServer(
+		kitot.TraceServer(tracer, "backup")(backupEndpoint(svc)),
+		decodeBackup,
+		encodeResponse,
+		opts...,
+	))
+
+	mux.Post("/restore", kithttp.NewServer(
+		kitot.TraceServer(tracer, "restore")(restoreEndpoint(svc)),
+		decodeRestore,
+		encodeResponse,
+		opts...,
+	))
+
+	return mux
+}
+
+func decodeBackup(_ context.Context, r *http.Request) (any, error) {
+	req := backupReq{token: apiutil.ExtractBearerToken(r)}
+
+	return req, nil
+}
+
+func decodeRestore(_ context.Context, r *http.Request) (any, error) {
+	if !strings.Contains(r.Header.Get("Content-Type"), apiutil.ContentTypeJSON) {
+		return nil, apiutil.ErrUnsupportedContentType
+	}
+
+	req := restoreReq{token: apiutil.ExtractBearerToken(r)}
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		return nil, errors.Wrap(apiutil.ErrMalformedEntity, err)
+	}
+
+	return req, nil
+}
+
+func encodeResponse(_ context.Context, w http.ResponseWriter, response any) error {
+	if ar, ok := response.(apiutil.Response); ok {
+		for k, v := range ar.Headers() {
+			w.Header().Set(k, v)
+		}
+		w.Header().Set("Content-Type", apiutil.ContentTypeJSON)
+		w.WriteHeader(ar.Code())
+
+		if ar.Empty() {
+			return nil
+		}
+	}
+
+	return json.NewEncoder(w).Encode(response)
+}
+
+func encodeError(_ context.Context, err error, w http.ResponseWriter) {
+	switch {
+	case errors.Contains(err, users.ErrPasswordFormat),
+		errors.Contains(err, errors.ErrInvalidPassword),
+		errors.Contains(err, users.ErrEmailVerificationExpired):
+		w.WriteHeader(http.StatusBadRequest)
+	case errors.Contains(err, uuid.ErrGeneratingID),
+		errors.Contains(err, users.ErrRecoveryToken):
+		w.WriteHeader(http.StatusInternalServerError)
+	case errors.Contains(err, users.ErrSelfRegisterDisabled):
+		w.WriteHeader(http.StatusForbidden)
+	default:
+		apiutil.EncodeError(err, w)
+	}
+
+	apiutil.WriteErrorResponse(err, w)
+}
