+++ users/api/http/doc.go
@@ -1,6 +1,5 @@
 // Copyright (c) Mainflux
 // SPDX-License-Identifier: Apache-2.0
 
-// Package api contains API-related concerns: endpoint definitions, middlewares
-// and all resource representations.
+// Package http contains implementation of users service HTTP API.
 package http
