+++ users/api/http/users/endpoint_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Mainflux
 // SPDX-License-Identifier: Apache-2.0
 
-package http_test
+package users_test
 
 import (
 	"context"
@@ -17,7 +17,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/MainfluxLabs/mainflux/auth"
 	"github.com/MainfluxLabs/mainflux/logger"
 	"github.com/MainfluxLabs/mainflux/pkg/apiutil"
 	"github.com/MainfluxLabs/mainflux/pkg/dbutil"
@@ -27,6 +26,7 @@ import (
 	"github.com/MainfluxLabs/mainflux/pkg/uuid"
 	"github.com/MainfluxLabs/mainflux/users"
 	httpapi "github.com/MainfluxLabs/mainflux/users/api/http"
+	svcusers "github.com/MainfluxLabs/mainflux/users/api/http/users"
 	usmocks "github.com/MainfluxLabs/mainflux/users/mocks"
 	"github.com/opentracing/opentracing-go/mocktracer"
 	"github.com/stretchr/testify/assert"
@@ -61,8 +61,7 @@ const (
 	invalidLimitData = `{"limit":210,"offset":0}`
 	invalidData      = `{"limit": "invalid"}`
 
-	inviteDuration     = 7 * 24 * time.Hour
-	inviteRedirectPath = "/register/invite"
+	inviteDuration = 7 * 24 * time.Hour
 )
 
 var (
@@ -108,11 +107,6 @@ var (
 
 	verificationsList = []users.EmailVerification{verification, duplicateVerification, expiredVerification}
 )
-
-type platformInviteReq struct {
-	Email        string `json:"email,omitempty"`
-	RedirectPath string `json:"redirect_path,omitempty"`
-}
 
 type selfRegisterReq struct {
 	User         users.User `json:"user,omitempty"`
@@ -838,7 +832,7 @@ func TestPasswordResetRequest(t *testing.T) {
 	expectedExisting := toJSON(struct {
 		Msg string `json:"msg"`
 	}{
-		httpapi.MailSent,
+		svcusers.MailSent,
 	})
 
 	cases := []struct {
@@ -1015,300 +1009,6 @@ func TestPasswordChange(t *testing.T) {
 
 		assert.Equal(t, tc.status, res.StatusCode, fmt.Sprintf("%s: expected status code %d got %d", tc.desc, tc.status, res.StatusCode))
 		assert.Equal(t, tc.res, token, fmt.Sprintf("%s: expected body %s got %s", tc.desc, tc.res, token))
-	}
-}
-
-func TestCreatePlatformInvite(t *testing.T) {
-	svc := newService()
-	ts := newServer(svc)
-	defer ts.Close()
-
-	client := ts.Client()
-
-	tokenAdmin, err := svc.Login(context.Background(), admin)
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s\n", err))
-
-	tokenRegular, err := svc.Login(context.Background(), user)
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s\n", err))
-
-	cases := []struct {
-		desc        string
-		req         string
-		contentType string
-		status      int
-		token       string
-	}{
-		{
-			"create platform invite",
-			toJSON(platformInviteReq{Email: "new@user.com", RedirectPath: inviteRedirectPath}),
-			contentType,
-			http.StatusCreated,
-			tokenAdmin,
-		},
-		{
-			"create platform invite as non-root-admin user",
-			toJSON(platformInviteReq{Email: "new@user.com", RedirectPath: inviteRedirectPath}),
-			contentType,
-			http.StatusForbidden,
-			tokenRegular,
-		},
-		{
-			"create platform invite with invalid auth token",
-			toJSON(platformInviteReq{Email: "new@user.com", RedirectPath: inviteRedirectPath}),
-			contentType,
-			http.StatusUnauthorized,
-			"invalid",
-		},
-		{
-			"create platform invite with empty redirect path",
-			toJSON(platformInviteReq{Email: "new@user.com", RedirectPath: ""}),
-			contentType,
-			http.StatusBadRequest,
-			tokenAdmin,
-		},
-		{
-			"create platform invite with empty email",
-			toJSON(platformInviteReq{Email: "", RedirectPath: inviteRedirectPath}),
-			contentType,
-			http.StatusBadRequest,
-			tokenAdmin,
-		},
-		{
-			"create platform invite with empty auth token",
-			toJSON(platformInviteReq{Email: "new@user.com", RedirectPath: inviteRedirectPath}),
-			contentType,
-			http.StatusUnauthorized,
-			"",
-		},
-		{
-			"create platform invite with empty request",
-			"",
-			contentType,
-			http.StatusBadRequest,
-			tokenAdmin,
-		},
-		{
-			"create platform invite with invalid request fromat",
-			"{,",
-			contentType,
-			http.StatusBadRequest,
-			tokenAdmin,
-		},
-	}
-
-	for _, tc := range cases {
-		req := testRequest{
-			client:      client,
-			method:      http.MethodPost,
-			url:         fmt.Sprintf("%s/invites", ts.URL),
-			contentType: tc.contentType,
-			body:        strings.NewReader(tc.req),
-			token:       tc.token,
-		}
-
-		res, err := req.make()
-		assert.Nil(t, err, fmt.Sprintf("%s: unexpected error %s", tc.desc, err))
-
-		assert.Equal(t, tc.status, res.StatusCode, fmt.Sprintf("%s: expected status code %d got %d", tc.desc, tc.status, res.StatusCode))
-	}
-}
-
-func TestViewPlatformInvite(t *testing.T) {
-	svc := newService()
-	ts := newServer(svc)
-	defer ts.Close()
-
-	client := ts.Client()
-
-	tokenAdmin, err := svc.Login(context.Background(), admin)
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s\n", err))
-
-	tokenRegular, err := svc.Login(context.Background(), user)
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s\n", err))
-
-	invite, err := svc.CreatePlatformInvite(context.Background(), tokenAdmin, inviteRedirectPath, "new@user.com", auth.OrgInvite{})
-	assert.Nil(t, err, fmt.Sprintf("Inviting platform member expected to succeed: %s\n", err))
-
-	cases := []struct {
-		desc     string
-		inviteID string
-		status   int
-		token    string
-	}{
-		{
-			"view platform invite",
-			invite.ID,
-			http.StatusOK,
-			tokenAdmin,
-		},
-		{
-			"view platform invite with invalid id",
-			"invalid-123",
-			http.StatusNotFound,
-			tokenAdmin,
-		},
-		{
-			"view platform invite as non-root-admin user",
-			invite.ID,
-			http.StatusForbidden,
-			tokenRegular,
-		},
-		{
-			"view platform invite with empty auth token",
-			invite.ID,
-			http.StatusUnauthorized,
-			"",
-		},
-		{
-			"view platform invite with invalid auth token",
-			invite.ID,
-			http.StatusUnauthorized,
-			"invalid123",
-		},
-	}
-
-	for _, tc := range cases {
-		req := testRequest{
-			client: client,
-			method: http.MethodGet,
-			url:    fmt.Sprintf("%s/invites/%s", ts.URL, tc.inviteID),
-			token:  tc.token,
-		}
-
-		res, err := req.make()
-		assert.Nil(t, err, fmt.Sprintf("%s: unexpected error %s", tc.desc, err))
-
-		assert.Equal(t, tc.status, res.StatusCode, fmt.Sprintf("%s: expected status code %d got %d", tc.desc, tc.status, res.StatusCode))
-	}
-}
-
-func TestListPlatformInvites(t *testing.T) {
-	svc := newService()
-	ts := newServer(svc)
-	defer ts.Close()
-
-	client := ts.Client()
-
-	tokenAdmin, err := svc.Login(context.Background(), admin)
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s\n", err))
-
-	tokenRegular, err := svc.Login(context.Background(), user)
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s\n", err))
-
-	_, err = svc.CreatePlatformInvite(context.Background(), tokenAdmin, inviteRedirectPath, "new@user.com", auth.OrgInvite{})
-	assert.Nil(t, err, fmt.Sprintf("Inviting platform member expected to succeed: %s\n", err))
-
-	_, err = svc.CreatePlatformInvite(context.Background(), tokenAdmin, inviteRedirectPath, "new1@user.com", auth.OrgInvite{})
-	assert.Nil(t, err, fmt.Sprintf("Inviting platform member expected to succeed: %s\n", err))
-
-	cases := []struct {
-		desc   string
-		status int
-		token  string
-	}{
-		{
-			"view platform invites",
-			http.StatusOK,
-			tokenAdmin,
-		},
-		{
-			"view platform invites as non-root-admin user",
-			http.StatusForbidden,
-			tokenRegular,
-		},
-		{
-			"view platform invites with empty auth token",
-			http.StatusUnauthorized,
-			"",
-		},
-		{
-			"view platform invites with invalid auth token",
-			http.StatusUnauthorized,
-			"invalid123",
-		},
-	}
-
-	for _, tc := range cases {
-		req := testRequest{
-			client: client,
-			method: http.MethodGet,
-			url:    fmt.Sprintf("%s/invites", ts.URL),
-			token:  tc.token,
-		}
-
-		res, err := req.make()
-		assert.Nil(t, err, fmt.Sprintf("%s: unexpected error %s", tc.desc, err))
-
-		assert.Equal(t, tc.status, res.StatusCode, fmt.Sprintf("%s: expected status code %d got %d", tc.desc, tc.status, res.StatusCode))
-	}
-}
-
-func TestRevokePlatformInvite(t *testing.T) {
-	svc := newService()
-	ts := newServer(svc)
-	defer ts.Close()
-
-	client := ts.Client()
-
-	tokenAdmin, err := svc.Login(context.Background(), admin)
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s\n", err))
-
-	tokenRegular, err := svc.Login(context.Background(), user)
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s\n", err))
-
-	invite, err := svc.CreatePlatformInvite(context.Background(), tokenAdmin, inviteRedirectPath, "new@user.com", auth.OrgInvite{})
-	assert.Nil(t, err, fmt.Sprintf("Inviting platform member expected to succeed: %s\n", err))
-
-	cases := []struct {
-		desc     string
-		inviteID string
-		status   int
-		token    string
-	}{
-		{
-			"revoke platform invite",
-			invite.ID,
-			http.StatusNoContent,
-			tokenAdmin,
-		},
-		{
-			"revoke platform invite as non-root-admin user",
-			invite.ID,
-			http.StatusForbidden,
-			tokenRegular,
-		},
-		{
-			"revoke platform invite with invalid invite id",
-			"invalid123",
-			http.StatusNotFound,
-			tokenAdmin,
-		},
-		{
-			"revoke platform invite with empty invalid auth token",
-			invite.ID,
-			http.StatusUnauthorized,
-			"invalid123",
-		},
-		{
-			"revoke platform invite with empty auth token",
-			invite.ID,
-			http.StatusUnauthorized,
-			"",
-		},
-	}
-
-	for _, tc := range cases {
-		req := testRequest{
-			client: client,
-			method: http.MethodDelete,
-			url:    fmt.Sprintf("%s/invites/%s", ts.URL, tc.inviteID),
-			token:  tc.token,
-		}
-
-		res, err := req.make()
-		assert.Nil(t, err, fmt.Sprintf("%s: unexpected error %s", tc.desc, err))
-
-		assert.Equal(t, tc.status, res.StatusCode, fmt.Sprintf("%s: expected status code %d got %d", tc.desc, tc.status, res.StatusCode))
 	}
 }
