+++ users/api/http/users/requests.go
@@ -1,12 +1,11 @@
 // Copyright (c) Mainflux
 // SPDX-License-Identifier: Apache-2.0
 
-package http
+package users
 
 import (
 	"regexp"
 
-	"github.com/MainfluxLabs/mainflux/auth"
 	"github.com/MainfluxLabs/mainflux/pkg/apiutil"
 	"github.com/MainfluxLabs/mainflux/pkg/errors"
 	"github.com/MainfluxLabs/mainflux/users"
@@ -43,28 +42,6 @@ func (req selfRegisterUserReq) validate() error {
 	}
 
 	return req.User.Validate(userPasswordRegex)
-}
-
-type registerByInviteReq struct {
-	User         users.User `json:"user"`
-	RedirectPath string     `json:"redirect_path"`
-	inviteID     string
-}
-
-func (req registerByInviteReq) validate() error {
-	if req.inviteID == "" {
-		return apiutil.ErrMissingInviteID
-	}
-
-	if err := req.User.Validate(userPasswordRegex); err != nil {
-		return err
-	}
-
-	if req.RedirectPath == "" {
-		return apiutil.ErrMissingRedirectPath
-	}
-
-	return nil
 }
 
 type verifyEmailReq struct {
@@ -237,108 +214,5 @@ func (req changeUserStatusReq) validate() error {
 	if req.id == "" {
 		return apiutil.ErrMissingUserID
 	}
-	return nil
-}
-
-type inviteReq struct {
-	token    string
-	inviteID string
-}
-
-func (req inviteReq) validate() error {
-	if req.token == "" {
-		return apiutil.ErrBearerToken
-	}
-
-	if req.inviteID == "" {
-		return apiutil.ErrMissingInviteID
-	}
-
-	return nil
-}
-
-type createPlatformInviteRequest struct {
-	token        string
-	Email        string             `json:"email,omitempty"`
-	OrgID        string             `json:"org_id"`
-	Role         string             `json:"role"`
-	GroupInvites []auth.GroupInvite `json:"group_invites"`
-	RedirectPath string             `json:"redirect_path,omitempty"`
-}
-
-func (req createPlatformInviteRequest) validate() error {
-	if req.token == "" {
-		return apiutil.ErrBearerToken
-	}
-
-	if req.Email == "" {
-		return apiutil.ErrMissingEmail
-	}
-
-	if req.RedirectPath == "" {
-		return apiutil.ErrMissingRedirectPath
-	}
-
-	if req.OrgID != "" && req.Role == "" {
-		return apiutil.ErrMissingRole
-	}
-
-	return nil
-}
-
-type listPlatformInvitesRequest struct {
-	token string
-	pm    users.PageMetadataInvites
-}
-
-func (req listPlatformInvitesRequest) validate() error {
-	if req.token == "" {
-		return apiutil.ErrBearerToken
-	}
-
-	if err := apiutil.ValidatePageMetadata(req.pm.PageMetadata, maxLimitSize, maxNameSize); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-type backupReq struct {
-	token string
-}
-
-func (req backupReq) validate() error {
-	if req.token == "" {
-		return apiutil.ErrBearerToken
-	}
-	return nil
-}
-
-type restoreUserReq struct {
-	ID       string         `json:"id"`
-	Email    string         `json:"email"`
-	Password string         `json:"password"`
-	Metadata map[string]any `json:"metadata"`
-	Status   string
-}
-type restoreReq struct {
-	token string
-	Users []restoreUserReq `json:"users"`
-	Admin restoreUserReq   `json:"admin"`
-}
-
-func (req restoreReq) validate() error {
-	if req.token == "" {
-		return apiutil.ErrBearerToken
-	}
-
-	if len(req.Users) == 0 {
-		return apiutil.ErrEmptyList
-	}
-
-	if req.Admin.ID == "" {
-		return apiutil.ErrMissingUserID
-	}
-
 	return nil
 }
